Implement support for floating point images as height maps and masks Implement support for 32-bit unsigned images as masks Show file selection dialog immediately on opening height map and mask import dialogs Make precision of height maps double instead of float Make height map implementations all read-only Make only necessary height map implementations serializable

+++ WorldPainter/WPCore/src/main/java/org/pepsoft/worldpainter/HeightMapTileFactory.java
@@ -82,7 +82,7 @@ public class HeightMapTileFactory extends AbstractTileFactory {
         return heightMap;
     }
     
-    public final float getBaseHeight() {
+    public final double getBaseHeight() {
         return heightMap.getBaseHeight();
     }
 
@@ -118,7 +118,7 @@ public class HeightMapTileFactory extends AbstractTileFactory {
             for (int x = 0; x < TILE_SIZE; x++) {
                 for (int y = 0; y < TILE_SIZE; y++) {
                     final int blockX = worldTileX + x, blockY = worldTileY + y;
-                    final float height = clamp(minHeight, heightMap.getHeight(blockX, blockY), maxZ);
+                    final float height = clamp(minHeight, (float) heightMap.getHeight(blockX, blockY), maxZ);
                     tile.setHeight(x, y, height);
                     tile.setWaterLevel(x, y, myWaterHeight);
                     if (floodWithLava) {
