+++ WorldPainter/WPCore/src/main/java/org/pepsoft/worldpainter/heightMaps/BicubicHeightMap.java
@@ -40,7 +40,7 @@ public class BicubicHeightMap extends DelegatingHeightMap {
     // HeightMap
 
     @Override
-    protected float doGetHeight(int x, int y) {
+    protected double doGetHeight(int x, int y) {
         if (repeat) {
             return children[0].getHeight(MathUtils.mod(x, width), MathUtils.mod(y, height));
         } else {
@@ -50,21 +50,21 @@ public class BicubicHeightMap extends DelegatingHeightMap {
 
     @SuppressWarnings("SuspiciousNameCombination") // Don't worry about it
     @Override
-    protected float doGetHeight(float x, float y) {
+    protected double doGetHeight(float x, float y) {
         x -= Math.signum(x) / 2;
         y -= Math.signum(y) / 2;
         final int xFloor = (int) Math.floor(x), yFloor = (int) Math.floor(y);
         final float xDelta = x - xFloor, yDelta = y - yFloor;
-        final float upperLeftValue = getExtHeight(xFloor, yFloor);
-        final float upperRightValue = getExtHeight(xFloor, yFloor + 1);
-        final float lowerLeftValue = getExtHeight(xFloor + 1, yFloor);
-        final float lowerRightValue = getExtHeight(xFloor + 1, yFloor + 1);
-        final float min = Math.min(Math.min(upperLeftValue, upperRightValue), Math.min(lowerLeftValue, lowerRightValue));
-        final float max = Math.max(Math.max(upperLeftValue, upperRightValue), Math.max(lowerLeftValue, lowerRightValue));
-        final float val1 = cubicInterpolate(getExtHeight(xFloor - 1, yFloor - 1), getExtHeight(xFloor - 1, yFloor), getExtHeight(xFloor - 1, yFloor + 1), getExtHeight(xFloor - 1, yFloor + 2), yDelta);
-        final float val2 = cubicInterpolate(getExtHeight(xFloor,     yFloor - 1), upperLeftValue,                   upperRightValue,                      getExtHeight(xFloor,     yFloor + 2), yDelta);
-        final float val3 = cubicInterpolate(getExtHeight(xFloor + 1, yFloor - 1), lowerLeftValue,                   lowerRightValue,                      getExtHeight(xFloor + 1, yFloor + 2), yDelta);
-        final float val4 = cubicInterpolate(getExtHeight(xFloor + 2, yFloor - 1), getExtHeight(xFloor + 2, yFloor), getExtHeight(xFloor + 2, yFloor + 1), getExtHeight(xFloor + 2, yFloor + 2), yDelta);
+        final double upperLeftValue = getExtHeight(xFloor, yFloor);
+        final double upperRightValue = getExtHeight(xFloor, yFloor + 1);
+        final double lowerLeftValue = getExtHeight(xFloor + 1, yFloor);
+        final double lowerRightValue = getExtHeight(xFloor + 1, yFloor + 1);
+        final double min = Math.min(Math.min(upperLeftValue, upperRightValue), Math.min(lowerLeftValue, lowerRightValue));
+        final double max = Math.max(Math.max(upperLeftValue, upperRightValue), Math.max(lowerLeftValue, lowerRightValue));
+        final double val1 = cubicInterpolate(getExtHeight(xFloor - 1, yFloor - 1), getExtHeight(xFloor - 1, yFloor), getExtHeight(xFloor - 1, yFloor + 1), getExtHeight(xFloor - 1, yFloor + 2), yDelta);
+        final double val2 = cubicInterpolate(getExtHeight(xFloor,     yFloor - 1), upperLeftValue,                   upperRightValue,                      getExtHeight(xFloor,     yFloor + 2), yDelta);
+        final double val3 = cubicInterpolate(getExtHeight(xFloor + 1, yFloor - 1), lowerLeftValue,                   lowerRightValue,                      getExtHeight(xFloor + 1, yFloor + 2), yDelta);
+        final double val4 = cubicInterpolate(getExtHeight(xFloor + 2, yFloor - 1), getExtHeight(xFloor + 2, yFloor), getExtHeight(xFloor + 2, yFloor + 1), getExtHeight(xFloor + 2, yFloor + 2), yDelta);
         // Constrain the value between the heights of the four corners, to try and mitigate haloing/ringing
         return clamp(min, cubicInterpolate(val1, val2, val3, val4, xDelta), max);
     }
@@ -74,7 +74,7 @@ public class BicubicHeightMap extends DelegatingHeightMap {
      * edge pixels of the image if it is non-repeating, to make the bicubic
      * interpolation work correctly around the edges.
      */
-    private float getExtHeight(float x, float y) {
+    private double getExtHeight(float x, float y) {
         if (repeat) {
             return children[0].getHeight(MathUtils.mod(x, width), MathUtils.mod(y, height));
         } else if ((extent == null) || extent.contains(x, y)) {
@@ -127,15 +127,15 @@ public class BicubicHeightMap extends DelegatingHeightMap {
     }
 
     @Override
-    public float[] getRange() {
+    public double[] getRange() {
         return children[0].getRange();
     }
 
     /**
      * Cubic interpolation using Catmull-Rom splines.
      */
-    private float cubicInterpolate(float y0, float y1, float y2, float y3, float μ) {
-        return y1 + 0.5f * μ * (y2 - y0 + μ * (2.0f * y0 - 5.0f * y1 + 4.0f * y2 - y3 + μ * (3.0f * (y1 - y2) + y3 - y0)));
+    private double cubicInterpolate(double y0, double y1, double y2, double y3, float μ) {
+        return y1 + 0.5 * μ * (y2 - y0 + μ * (2.0 * y0 - 5.0 * y1 + 4.0 * y2 - y3 + μ * (3.0 * (y1 - y2) + y3 - y0)));
     }
 
     private final int width, height;
