+++ WorldPainter/WPCore/src/main/java/org/pepsoft/worldpainter/heightMaps/MandelbrotHeightMap.java
@@ -7,7 +7,7 @@ import javax.swing.*;
  */
 public class MandelbrotHeightMap extends AbstractHeightMap {
     @Override
-    public float getHeight(float x0, float y0) {
+    public double getHeight(float x0, float y0) {
         float x = 0.0f, y = 0.0f;
         int iteration = 0;
         while ((x * x + y * y < 4) && (iteration < MAX_ITERATIONS)) {
@@ -25,10 +25,10 @@ public class MandelbrotHeightMap extends AbstractHeightMap {
     }
 
     @Override
-    public float[] getRange() {
+    public double[] getRange() {
         return RANGE;
     }
 
     private static final int MAX_ITERATIONS = 255;
-    private static final float[] RANGE = {0.0f, MAX_ITERATIONS};
+    private static final double[] RANGE = {0.0, MAX_ITERATIONS};
 }
